Made Guild Settings Into A Cog
